web/requirements.txt
airtable==0.4.8
discord==2.0.0
discord.py==2.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0